extracting specialityItem and groupItem components to their own files
