+++ client/src/components/utils/modals/components/addClass.jsx
@@ -8,148 +8,8 @@ import ModalContext from "../../../../datamanager/contexts/modalContext"
 import Select from "../../inputs/select"
 import { BsFillPlusCircleFill, BsX } from 'react-icons/bs'
 import reducer, { initialState } from '../reducers/classReducer'
-
-const GroupItem = ({ data: { id, name, capacity }, onDeleteGroup }) => {
-  return (
-    <Box
-      className={classStyles.groupItem}
-    >
-      <span className={classStyles.groupItemTitle}>{ name }</span>
-
-      {
-        id > 1 && (
-          <Box
-            sx={{
-              position: 'absolute',
-              top: 0,
-              right: 5,
-              "&:hover": {
-                cursor: "pointer"
-              }
-            }}
-            onClick={() => onDeleteGroup(id)}
-          >
-            <BsX 
-              color="#828282"
-              size={25}
-            />
-          </Box>
-        )
-      }
-
-      <Input 
-        placeholder="capacité"
-        type="number"
-        fullWidth
-        sx={{
-          mt: 1
-        }}
-        value={capacity}
-      />
-    </Box>
-  )
-}
-
-const SpecialityItem = ({ 
-  data: { 
-    id, 
-    value, 
-    capacity, 
-    groups 
-  }, 
-  index, 
-  onDeleteSpeciality,
-  onUpdateSpeciality,
-  onAddSpecialityGroup,
-  onDeleteSpecialityGroup
-}) => {
-  return (
-    <Box className={classStyles.groupContainer}>
-      <span 
-        className={classStyles.groupItemTitle} 
-        style={{ 
-          display: "block",
-          marginBottom: 15 
-        }}
-      >
-        { "specialite " + index }
-      </span>
-
-      <Box
-        sx={{
-          position: 'absolute',
-          top: 5,
-          right: 5,
-          "&:hover": {
-            cursor: "pointer"
-          }
-        }}
-        onClick={() => onDeleteSpeciality(id)}
-      >
-        <BsX 
-          color="#828282"
-          size={25}
-        />
-      </Box>
-
-      <Select 
-        label="nom"
-        options={[
-          { value: 0, label: "Geni Logiciel" },
-          { value: 1, label: "Reseau" }
-        ]}
-        value={value}
-        fullWidth
-        onGetValue={(value) => onUpdateSpeciality(id, "value", value)}
-      />
-      <Input 
-        placeholder="capacité"
-        type="number"
-        fullWidth
-        sx={{
-          mt: 1
-        }}
-        value={capacity}
-        onChange={(e) => onUpdateSpeciality(id, "capacity", e.target.value)}
-      />
-
-      <span className={classStyles.groupLabel}>Groupes de la spécialité</span>
-
-      <Box 
-        className={`
-          ${classStyles.groupContainer}
-          ${classStyles.specialityGroupContainer}
-        `}
-      >
-        {
-          groups.map((item) => {
-            return (
-              <GroupItem 
-                key={item.id}
-                data={item}
-                onDeleteGroup={(idGroup) => onDeleteSpecialityGroup(id, idGroup)}
-              />
-            )
-          })
-        }
-
-        <Button 
-          text="nouveau groupe"
-          variant="outlined"
-          bgColor="#ff8500"
-          fontSize={12}
-          rounded
-          onClick={() => onAddSpecialityGroup(id)}
-        >
-          <BsFillPlusCircleFill 
-            size={15}
-            color="#ff8500"
-          />
-        </Button>
-      </Box>
-    </Box>
-  )
-}
+import GroupItem from "./subComponents/groupItem"
+import SpecialityItem from './subComponents/specialityItem'
 
 const AddClassModalContent = () => {
   // Get global state
